backend/requirements.txt
asgiref==3.2.10
attrs==19.3.0
Django==3.1
django-rest-framework==0.1.0
djangorestframework==3.11.1
importlib-metadata==1.7.0
jsonschema==3.2.0
Pillow==7.2.0
psycopg2==2.8.5
pyrsistent==0.16.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
zipp==3.1.0